create search.js and tags.js file, move logic from page-load.js
