+++ src/page-load.js
@@ -1,81 +1,4 @@
-import weatherData from './weather'
-
-let getImage = (code) => {
-  let url = `http://openweathermap.org/img/wn/${code}@2x.png`
-  return url
-}
-
-let getTags = () => {
-  let cityName = document.querySelector('.city');
-  let img = document.querySelector('.weather-img');
-  let temperature = document.querySelector('.temperature');
-  let condition = document.querySelector('.condition');
-  let description = document.querySelector('.description');
-  let feelsLike = document.querySelector('.feels-like');
-  let minTemp = document.querySelector('.min-temp');
-  let maxTemp = document.querySelector('.max-temp');
-
-  return { cityName, img, temperature, condition, description, feelsLike, minTemp, maxTemp }
-}
-
-let setUnitString = (unit) => {
-  let string;
-
-  switch (unit) {
-    case 'imperial':
-      string = '°F'
-      break;
-    case 'standard':
-      string = '°K'
-      break;
-    default:
-      string = '°C'
-  }
-
-  return string
-}
-
-let setTags = (temperatures, weather, unit) => {
-  let tags = getTags();
-
-  let unitTemp = setUnitString(unit);
-
-  tags.cityName.textContent = weatherData.getName();
-  tags.temperature.textContent = `${parseInt(temperatures.temp)} ${unitTemp}`
-  tags.img.src = getImage(weather.icon)
-  tags.condition.textContent = weather.main;
-  tags.description.textContent = weather.description;
-  tags.feelsLike.textContent = `Feel: ${parseInt(temperatures.feels_like)} ${unitTemp}`;
-  tags.minTemp.textContent = `Min: ${parseInt(temperatures.temp_min)} ${unitTemp}`;
-  tags.maxTemp.textContent = `Max: ${parseInt(temperatures.temp_max)} ${unitTemp}`;
-}
-
-let ubicationSuccess = async function(data) {
-  await weatherData.getWeatherByUbication(41, 12);
-  let temperatures = weatherData.getTemperatures();
-  let weatherInfo = weatherData.getWeatherInfo();
-  setTags(temperatures, weatherInfo);
-}
-
-let getUbication = () => {
-  navigator.geolocation.getCurrentPosition(async function(data) { await ubicationSuccess(data) }, function(err) { alert(`Please allow location access`) });
-}
-
-let searchByCity = async function(cityName, unit) {
-  try {
-    await weatherData.getWeatherByCity(cityName, unit);
-    let temperatures = weatherData.getTemperatures();
-    let weatherInfo = weatherData.getWeatherInfo();
-    setTags(temperatures, weatherInfo, unit);
-  }
-  catch {
-    let notice = document.createElement('div');
-    notice.textContent = 'This is\'nt a valid country'
-
-    document.body.appendChild(notice)
-  }
-}
-
+import searchMod from './search'
 
 let pageLoad = () => {
   let container = document.querySelector('.container');
@@ -83,7 +6,7 @@ let pageLoad = () => {
   let unitSelect = document.querySelector('.unit');
   let searchBtn = document.querySelector('.search-btn');
 
-  searchBtn.onclick = () => {searchByCity(cityField.value || 'New york', unitSelect.value)}
+  searchBtn.onclick = () => {searchMod.searchByCity(cityField.value || 'New york', unitSelect.value)}
 
   let card = document.createElement('div');
   card.classList.add('card');
@@ -107,7 +30,7 @@ let pageLoad = () => {
   maxTemp.classList.add('max-temp')
 
   let ubication = () => { 
-    getUbication();
+    searchMod.searchByUbication();
   }
 
   ubication();
