+++ src/style.css
@@ -1,0 +1,33 @@
+* {
+  font-family: Arial, Helvetica, sans-serif
+}
+
+.wrapper {
+  height: 100vh;
+}
+
+.d-flex {
+  display: flex;
+}
+
+.justify-content-center {
+  justify-content: center;
+}
+
+.aling-items-center {
+  align-items: center;
+}
+
+.flex-column {
+  flex-direction: column;
+}
+
+.search {
+  position: fixed;
+  top: 20px;
+}
+
+.city {
+  font-weight: bold;
+  font-size: 1.5rem;
+}
